apps/api/pyproject.toml
[project]
name = "genpos-api"
version = "0.1.0"
description = "GenPos API - XiaoHongShu AI Ads Workspace Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "boto3>=1.36.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "temporalio>=1.9.0",
    "openai>=1.60.0",
    "python-multipart>=0.0.18",
    "Pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]
